Cambios en Busqueda
ahora busca dentro de los articulos

+++ SistemaOrdenes/Detalles.cs
@@ -8,7 +8,7 @@ namespace SistemaOrdenes
     class Detalles
     {
         public int id_orden;
-        public int cantidad;
+        public double cantidad;
         public String descripcion;
         public float punitario;
         public String moneda;
